NAE-335 - Variable arc definition rework
- implemented changes from pull request
- fixed VariableArcsTest
- fixed bug where reagular arcs did not return tokens after cancel in TaskService method returnTokens

+++ src/main/java/com/netgrif/workflow/workflow/domain/DataField.java
@@ -191,6 +191,6 @@ public class DataField implements Referencable {
 
     @Override
     public int getMultiplicity() {
-        return Integer.parseInt(String.valueOf(value));
+        return (int) Double.parseDouble(String.valueOf(value));
     }
 }
